filter.subsample: Move subsampling functions here
